Refactored Controller code into Controller Manager class
